IGNITE-28901 Split GridEventStorageMessage into a request and a response
