+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -735,7 +735,6 @@ org.apache.ignite.internal.managers.encryption.GridEncryptionManager$EmptyResult
 org.apache.ignite.internal.managers.encryption.GridEncryptionManager$MasterKeyChangeRequest
 org.apache.ignite.internal.managers.encryption.NodeEncryptionKeys
 org.apache.ignite.internal.managers.encryption.GroupKeyEncrypted
-org.apache.ignite.internal.managers.eventstorage.GridEventStorageMessage
 org.apache.ignite.internal.managers.indexing.GridIndexingManager$1
 org.apache.ignite.internal.managers.loadbalancer.GridLoadBalancerAdapter
 org.apache.ignite.internal.managers.loadbalancer.GridLoadBalancerManager$1
